Monthly Calendar moved to Navigation Drawer
