+++ .idea/misc.xml
@@ -14,6 +14,7 @@
         <entry key="..\:/Users/TIRTH JOSHI/AndroidStudioProjects/US_2_0/app/src/main/res/layout/activity_main.xml" value="0.22" />
         <entry key="..\:/Users/User/Desktop/Hack/app/src/main/res/layout/activity_event_edit.xml" value="0.10144927536231885" />
         <entry key="..\:/Users/User/Desktop/Hack/app/src/main/res/layout/activity_landing_page.xml" value="0.10144927536231885" />
+        <entry key="..\:/Users/User/Desktop/Hack/app/src/main/res/layout/activity_month_view.xml" value="0.10144927536231885" />
         <entry key="..\:/Users/User/Desktop/Hack/app/src/main/res/layout/activity_week_view.xml" value="0.10144927536231885" />
       </map>
     </option>
