+++ app/src/main/res/layout/activity_month_view.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    tools:context=".MonthViewActivity"
     android:orientation="vertical"
-    tools:context=".WeekViewActivity">
+    android:background="@color/gray">
+
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -24,7 +25,7 @@
             android:textColor="@color/white"
             android:background="@null"
             android:textStyle="bold"
-            android:onClick="previousWeekAction"
+            android:onClick="previousMonthAction"
             android:textSize="20sp"/>
 
         <TextView
@@ -34,8 +35,9 @@
             android:layout_height="wrap_content"
             android:text="Feb 2021"
             android:textSize="20sp"
+            android:backgroundTint="@color/gray"
             android:textAlignment="center"
-            android:textColor="@color/purple_200"/>
+            android:textColor="@color/white"/>
 
         <Button
             android:layout_width="0dp"
@@ -46,7 +48,7 @@
             android:textColor="@color/white"
             android:background="@null"
             android:textStyle="bold"
-            android:onClick="nextWeekAction"
+            android:onClick="nextMonthAction"
             android:textSize="20sp"/>
 
     </LinearLayout>
@@ -123,27 +125,17 @@
     </LinearLayout>
 
 
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:background="#000000"/>
+
+
+
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/calendarRecyclerView"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"/>
-
-    <ListView
-        android:id="@+id/eventListView"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="5"/>
-
-
+        android:layout_height="match_parent"/>
 </LinearLayout>
-
-
-
-
-
-
-
-
-
